+++ CHANGELOG.md
@@ -4,6 +4,8 @@ CHANGELOG](http://keepachangelog.com/) for how to update this file. This project
 adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Improved AWS Lambda integration with `hb_lambda` decorator
 
 ## [4.11.0] - 2022-02-15
 ### Fixed